requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
websocket-client>=1.7.0
pymongo>=4.7.0
qdrant-client>=1.7.0

